Send ships into a death spiral when player disconnects
Had to fix up the monitoring logic, was having issues because multiple monitors
results in multiple down messages. Also monitoring in `init` didn't make sense
because if the player reconnects, the monitored pid needs to change.

Also commit commented out versions of some of the debugging code that I was using